Refactoring: GridCreatingConditionTriangle::AddHolePolygonCommand implementation is separated from gridcreatingconditiontriangle.cpp.
